Affirmative action
For the most part affirmative action is complete but I wanted to see if
adam or kyle could clean it up because it does not print on the middle
of the page where its suppose to be. It prints below for some reason.
Also I need to figure out how to make the sub tab affirmative action
bigger because the button is too small. The words run off the button.